+++ test/model.test.js
@@ -1122,16 +1122,15 @@ module.exports = {
 
   // GH-319
   'save callback should only execute once regardless of number of failed validations': function () {
-    mongoose.model('CallbackFiresOnceValidation', new Schema({
+    var db = start()
+
+    var D = db.model('CallbackFiresOnceValidation', new Schema({
         username: { type: String, validate: /^[a-z]{6}$/i }
       , email: { type: String, validate: /^[a-z]{6}$/i }
       , password: { type: String, validate: /^[a-z]{6}$/i }
     }));
 
-    var db = start()
-      , CallbackFiresOnceValidation = db.model('CallbackFiresOnceValidation');
-
-    var post = new CallbackFiresOnceValidation({
+    var post = new D({
         username: "nope"
       , email: "too"
       , password: "short"
@@ -1140,6 +1139,7 @@ module.exports = {
     var timesCalled = 0;
 
     post.save(function (err) {
+      db.close();
       err.should.be.an.instanceof(MongooseError);
       err.should.be.an.instanceof(ValidationError);
 
@@ -1161,7 +1161,6 @@ module.exports = {
       post.errors.email.message.should.eql('Validator failed for path email');
       post.errors.username.message.should.eql('Validator failed for path username');
 
-      db.close();
     });
   },
 
@@ -4613,111 +4612,6 @@ module.exports = {
     });
   },
 
-  'strict mode': function(){
-    var db = start();
-
-    var lax = new Schema({
-        ts  : { type: Date, default: Date.now }
-      , content: String
-    });
-
-    var strict = new Schema({
-        ts  : { type: Date, default: Date.now }
-      , content: String
-    }, { strict: true });
-
-    var Lax = db.model('Lax', lax);
-    var Strict = db.model('Strict', strict);
-
-    var l = new Lax({content: 'sample', rouge: 'data'});
-    l._strictMode.should.be.false;
-    l = l.toObject();
-    l.content.should.equal('sample')
-    l.rouge.should.equal('data');
-    should.exist(l.rouge);
-
-    var s = new Strict({content: 'sample', rouge: 'data'});
-    s._strictMode.should.be.true;
-    s = s.toObject();
-    s.should.have.property('ts');
-    s.content.should.equal('sample');
-    s.should.not.have.property('rouge');
-    should.not.exist(s.rouge);
-
-    // instance override
-    var instance = new Lax({content: 'sample', rouge: 'data'}, true);
-    instance._strictMode.should.be.true;
-    instance = instance.toObject();
-    instance.content.should.equal('sample')
-    should.not.exist(instance.rouge);
-    instance.should.have.property('ts')
-
-    // hydrate works as normal, but supports the schema level flag.
-    var s2 = new Strict({content: 'sample', rouge: 'data'}, false);
-    s2._strictMode.should.be.false;
-    s2 = s2.toObject();
-    s2.should.have.property('ts')
-    s2.content.should.equal('sample');
-    s2.should.have.property('rouge');
-    should.exist(s2.rouge);
-
-    // testing init
-    var s3 = new Strict();
-    s3.init({content: 'sample', rouge: 'data'});
-    var s3obj = s3.toObject();
-    s3.content.should.equal('sample');
-    s3.should.not.have.property('rouge');
-    should.not.exist(s3.rouge);
-
-    // strict on create
-    Strict.create({content: 'sample2', rouge: 'data'}, function(err, doc){
-      doc.content.should.equal('sample2');
-      doc.should.not.have.property('rouge');
-      should.not.exist(doc.rouge);
-      db.close();
-    });
-  },
-
-  'strict mode virtuals': function () {
-    var db = start();
-
-    var getCount = 0
-      , setCount = 0;
-
-    var strictSchema = new Schema({
-        email: String
-      , prop: String
-    }, {strict: true});
-
-    strictSchema
-    .virtual('myvirtual')
-    .get(function() {
-      getCount++;
-      return 'ok';
-    })
-    .set(function(v) {
-      setCount++;
-      this.prop = v;
-    });
-
-    var StrictModel = db.model('StrictVirtual', strictSchema);
-
-    var strictInstance = new StrictModel({
-        email: 'hunter@skookum.com'
-      , myvirtual: 'test'
-    });
-
-    db.close();
-    getCount.should.equal(0);
-    setCount.should.equal(1);
-
-    strictInstance.myvirtual = 'anotherone';
-    var myvirtual = strictInstance.myvirtual;
-
-    getCount.should.equal(1);
-    setCount.should.equal(2);
-  },
-
   'should not throw range error when using Number _id and saving existing doc': function () {
     var db =start();
 
@@ -4737,6 +4631,23 @@ module.exports = {
           db.close();
           should.strictEqual(null, err);
         });
+      });
+    });
+  },
+
+  'number of affected docs should be returned when saving': function () {
+    var db = start()
+    var schema = new Schema({ name: String });
+    var S = db.model('AffectedDocsAreReturned', schema);
+    var s = new S({ name: 'aaron' });
+    s.save(function (err, doc, affected) {
+      should.strictEqual(null, err);
+      affected.should.equal(1);
+      s.name = 'heckmanananananana';
+      s.save(function (err, doc, affected) {
+        db.close();
+        should.strictEqual(null, err);
+        affected.should.equal(1);
       });
     });
   }